Revert "Revert "Remove legacy POA indexing code (#4449)" (#4566)"
This reverts commit 97bee1ab13d4aa9e5e49254dade7b83b217205b9.
