+++ discovery-provider/src/utils/constants.py
@@ -1028,21 +1028,9 @@ default_lengths = {
 
 
 class CONTRACT_TYPES(Enum):
-    USER_FACTORY = "user_factory"
-    TRACK_FACTORY = "track_factory"
-    SOCIAL_FEATURE_FACTORY = "social_feature_factory"
-    PLAYLIST_FACTORY = "playlist_factory"
-    USER_LIBRARY_FACTORY = "user_library_factory"
-    USER_REPLICA_SET_MANAGER = "user_replica_set_manager"
     ENTITY_MANAGER = "entity_manager"
 
 
 CONTRACT_NAMES_ON_CHAIN = {
-    CONTRACT_TYPES.USER_FACTORY: "UserFactory",
-    CONTRACT_TYPES.TRACK_FACTORY: "TrackFactory",
-    CONTRACT_TYPES.SOCIAL_FEATURE_FACTORY: "SocialFeatureFactory",
-    CONTRACT_TYPES.PLAYLIST_FACTORY: "PlaylistFactory",
-    CONTRACT_TYPES.USER_LIBRARY_FACTORY: "UserLibraryFactory",
-    CONTRACT_TYPES.USER_REPLICA_SET_MANAGER: "UserReplicaSetManager",
     CONTRACT_TYPES.ENTITY_MANAGER: "EntityManager",
 }
