+++ discovery-provider/src/utils/prometheus_metric.py
@@ -103,13 +103,11 @@ class PrometheusMetricNames:
     INDEX_BLOCKS_DURATION_SECONDS = "index_blocks_duration_seconds"
     INDEX_METRICS_DURATION_SECONDS = "index_metrics_duration_seconds"
     INDEX_TRENDING_DURATION_SECONDS = "index_trending_duration_seconds"
-    TRACK_STATE_UPDATE_DURATION_SECONDS = "track_state_update_duration_seconds"
     UPDATE_AGGREGATE_TABLE_DURATION_SECONDS = "update_aggregate_table_duration_seconds"
     UPDATE_TRACK_IS_AVAILABLE_DURATION_SECONDS = (
         "update_track_is_available_duration_seconds"
     )
     UPDATE_TRENDING_VIEW_DURATION_SECONDS = "update_trending_view_duration_seconds"
-    USER_STATE_UPDATE_DURATION_SECONDS = "user_state_update_duration_seconds"
     ENTITY_MANAGER_UPDATE_CHANGED_LATEST = "entity_manager_update_changed_latest"
     ENTITY_MANAGER_UPDATE_DURATION_SECONDS = "entity_manager_update_duration_seconds"
     ENTITY_MANAGER_UPDATE_ERRORS = "entity_manager_update_errors"
@@ -196,11 +194,6 @@ PrometheusRegistry = {
         f"{METRIC_PREFIX}_{PrometheusMetricNames.INDEX_TRENDING_DURATION_SECONDS}",
         "Runtimes for src.task.index_trending:index_trending()",
     ),
-    PrometheusMetricNames.TRACK_STATE_UPDATE_DURATION_SECONDS: Histogram(
-        f"{METRIC_PREFIX}_{PrometheusMetricNames.TRACK_STATE_UPDATE_DURATION_SECONDS}",
-        "Runtimes for src.task.tracks:track_state_update()",
-        ("scope",),
-    ),
     PrometheusMetricNames.UPDATE_AGGREGATE_TABLE_DURATION_SECONDS: Histogram(
         f"{METRIC_PREFIX}_{PrometheusMetricNames.UPDATE_AGGREGATE_TABLE_DURATION_SECONDS}",
         "Runtimes for src.task.aggregates:update_aggregate_table()",
@@ -218,11 +211,6 @@ PrometheusRegistry = {
         f"{METRIC_PREFIX}_{PrometheusMetricNames.UPDATE_TRENDING_VIEW_DURATION_SECONDS}",
         "Runtimes for src.task.index_trending:update_view()",
         ("mat_view_name",),
-    ),
-    PrometheusMetricNames.USER_STATE_UPDATE_DURATION_SECONDS: Histogram(
-        f"{METRIC_PREFIX}_{PrometheusMetricNames.USER_STATE_UPDATE_DURATION_SECONDS}",
-        "Runtimes for src.task.users:user_state_update()",
-        ("scope",),
     ),
     PrometheusMetricNames.ENTITY_MANAGER_UPDATE_CHANGED_LATEST: Histogram(
         f"{METRIC_PREFIX}_{PrometheusMetricNames.ENTITY_MANAGER_UPDATE_CHANGED_LATEST}",
